% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gertest.R
\name{ger_test}
\alias{ger_test}
\title{Conduct a range of germination rate comparison tests.}
\usage{
ger_test(gdata = data, method = "LRT", p_adjust_mtd = "holm", ctr_lv = 0.05)
}
\arguments{
\item{gdata}{A data-frame containing the germination data
('gdata,' is expected to be a data frame containing three variables:
'treatment,' 'num_seed,' and 'ger_seed'. The 'treatment' variable is a factor
variable representing the treatment level, 'num_seed' should indicate
the number of seeds used in each iteration of the treatment,
and 'ger_seed' should indicate the number of seeds that germinated.)}

\item{method}{A statistical method for germination rate comparison
(The options for method are "LRT" (likelihood ratio test), "ANOVA"
(analysis of variance), "AS-ANOVA" (ANOVA with arcsine square root transformed response),
"logistic" (logistic regression), or "KW" (Kruskal-Wallis test).}

\item{p_adjust_mtd}{P-value correction method for multiple comparison
The options for the correction method are
c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none").
For ANOVA with Tukey's HSD, the method should be "single-step"}

\item{ctr_lv}{significance level for overall test and control level for
multiple testing (we assume that both levels are the same)}
}
\value{
p_value the resulting P-value from the test

cld compact letter display from the multiple testing
(When the p-value is smaller than the specified significance level,
the cld becomes NULL)
}
\description{
The function executes a germination rate comparison test based on
the specified method. If the difference in germination rate comparison is
statistically significant at the designated significance level, it generates a compact
letter display that groups treatments with similar germination rates,
utilizing the specified multiple comparison method.
}
\examples{
\dontrun{
# Data import from a csv file
gerdata<-read.csv('D:\\\\nandina_data.csv', header=TRUE)
ger_test(gdata=gerdata,method="ANOVA",method="single-step")
}

# Data generation
data(nandina)
ger_test(gdata=nandina,method="LRT")
}
