% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerstat-package.R
\docType{package}
\name{gerstat-package}
\alias{gerstat}
\alias{gerstat-package}
\title{gerstat: Offers a diverse range of statistical tests to compare germination rates}
\description{
A toolbox to easily perform germination rate comparisons. Provides the P-value for a test for the presence of a difference in germination rate. It also provides a compact letter display based on the results of a multiple comparison test for germination rate differences.
}
\author{
\strong{Maintainer}: Hohsuk Noh \email{hsnoh@sookmyung.ac.kr}

}
\keyword{internal}
