% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nandina}
\alias{nandina}
\title{Germination Data of Heavenly Bamboo}
\format{
\subsection{\code{nandina}}{

A data frame with 12 rows and 4 columns:
\describe{
\item{treatment}{treatment level}
\item{num_seed}{number of tested seeds}
\item{ger_seed}{number of germinated seeds}
\item{rep}{replicate number}
}
}
}
\usage{
nandina
}
\description{
A subset of data from Rhie et al. (2016)
(Non-deep simple morphophysiological dormancy in seeds of heavenly bamboo)
}
\keyword{datasets}
